+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-v0.6.7 (Jul 07 2023)
+v0.6.8 (Aug 02 2023)
 ===============================================================================
 ## Added
  * `report` now take a `LogValue` or a list of `LogValue`s.
 
 ## Changed
- * It will now bulk data, if a list af `LogValue`s is given.
+ * It will now bulk report data, if a list af `LogValue`s is given.
+ * Bulk report data will now be sorted.
 
 ## Fixed
  * Handle a rare error case from the server.